fix: multi-threaded handle the received message
